+++ test/unit-tests/banners-summary.test.tsx
@@ -1,14 +1,10 @@
 import {VersionParts} from "@/banners/types";
 import {
     BannerSummary,
-    getAvgBannerGapInterval,
-    getAvgDayInterval,
-    getAvgPatchGapInterval,
     getBannerGap,
     getPatchGap,
     getBannersSinceLastCountSummary,
     getDaysSinceRunCountSummary,
-    getResourceSummaries,
     getPatchesSinceLastCountSummary,
     getRunsCountSummary, getAverageDaysInBetween, getAverageBannersInBetween, getAveragePatchesInBetween
 } from "@/banners/summary";
@@ -114,120 +110,6 @@ describe('getBannerGap', () => {
             .toEqual(-1)
     });
 })
-describe('getAvgPatchGapInterval', () => {
-    it('should get the avg patch gap for a character', async () => {
-        expect(getAvgPatchGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Hu Tao"]))
-            .toEqual(6)
-
-        expect(getAvgPatchGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Venti"]))
-            .toEqual(5.7)
-
-        expect(getAvgPatchGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Yoimiya"]))
-            .toEqual(5.5)
-    });
-
-    it('should return -1 when character has only ran once', async () => {
-        expect(getAvgPatchGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Fake"]))
-            .toEqual(-1)
-    });
-})
-
-describe('getAvgBannerGapInterval', () => {
-    it('should get the avg banner gap for a character', async () => {
-        expect(getAvgBannerGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Hu Tao"]))
-            .toEqual(11)
-
-        expect(getAvgBannerGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Venti"]))
-            .toEqual(10.7)
-
-        expect(getAvgBannerGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Yoimiya"]))
-            .toEqual(9.5)
-    });
-
-    it('should return -1 when character has only ran once', async () => {
-        expect(getAvgBannerGapInterval(VersionPartsDummyData, BannerSummariesDummyData["Fake"]))
-            .toEqual(-1)
-    });
-})
-
-describe('getAvgDayInterval', () => {
-    it('should get the avg day interval for a character', async () => {
-        expect(getAvgDayInterval(BannerSummariesDummyData["Hu Tao"]))
-            .toEqual(231)
-
-        expect(getAvgDayInterval(BannerSummariesDummyData["Venti"]))
-            .toEqual(223.3)
-
-        expect(getAvgDayInterval(BannerSummariesDummyData["Yoimiya"]))
-            .toEqual(203.5)
-    });
-
-    it('should return -1 when character has only ran once', async () => {
-        expect(getAvgDayInterval(BannerSummariesDummyData["Fake"]))
-            .toEqual(-1)
-    });
-})
-
-
-describe('getResourceSummaries', () => {
-    it('should get resource summary correctly', async () => {
-        expect(_.orderBy(
-            getResourceSummaries(VersionPartsDummyData, BannerSummariesDummyData, dayjs.utc("2023-01-03")),
-            (b) => b.name,
-            'asc'
-        ))
-            .toEqual(_.orderBy([
-                    {
-                        "name": "Fake",
-                        "image": "Fake",
-                        "avgBannerGapInterval": -1,
-                        "avgDaysInterval": -1,
-                        "avgPatchGapInterval": -1,
-                        "daysSinceLastRun": 490,
-                        "bannersSinceLastRun": 22,
-                        "patchesSinceLastRun": 11,
-                        "runs": 1
-                    },
-                    {
-                        "name": "Hu Tao",
-                        "image": "Hu-Tao",
-                        "avgBannerGapInterval": 11,
-                        "avgDaysInterval": 231,
-                        "avgPatchGapInterval": 6,
-                        "daysSinceLastRun": 406,
-                        "bannersSinceLastRun": 18,
-                        "patchesSinceLastRun": 9,
-                        "runs": 2
-                    },
-                    {
-                        "name": "Venti",
-                        "image": "Venti",
-                        "avgBannerGapInterval": 10.7,
-                        "avgDaysInterval": 223.3,
-                        "avgPatchGapInterval": 5.7,
-                        "daysSinceLastRun": 81,
-                        "bannersSinceLastRun": 3,
-                        "patchesSinceLastRun": 1,
-                        "runs": 4
-                    },
-                    {
-                        "name": "Yoimiya",
-                        "image": "Yoimiya",
-                        "avgBannerGapInterval": 9.5,
-                        "avgDaysInterval": 203.5,
-                        "avgPatchGapInterval": 5.5,
-                        "daysSinceLastRun": 46,
-                        "bannersSinceLastRun": 1,
-                        "patchesSinceLastRun": 0,
-                        "runs": 3
-                    },
-                ],
-                (b) => b.name,
-                'asc',
-            ))
-    });
-})
-
 
 describe('getDaysSinceRunCountSummary', () => {
     it('should return the number of days since last run', () => {
